{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["Name: {{cookiecutter.email.split('@')[0]}} & Email: {{cookiecutter.email}}"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8, <3.9"
pandas = ">=2.0.0"
numpy = "1.24.4"
pyarrow = "^15.0.2"
scikit-learn = "0.22.1"
pip = "^24.0"
azureml-core = "1.53.0"
azure-keyvault-secrets = "4.6.0"
matplotlib = "3.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pre-commit = "3.5.0"
pylint = "^3.1.0"
cookiecutter = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
